+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/project/Configurations.kt
@@ -38,8 +38,7 @@ internal enum class WorkspaceFolders(val label: String) {
 }
 
 
-internal enum class DiagnosticMode(val value: String?, val label: String) {
-    DEFAULT(null, message("configurations.diagnosticMode.default")),
+internal enum class DiagnosticMode(val value: String, val label: String) {
     OPEN_FILES_ONLY("openFilesOnly", message("configurations.diagnosticMode.openFilesOnly")),
     WORKSPACE("workspace", message("configurations.diagnosticMode.workspace"));
 }
@@ -50,5 +49,6 @@ internal class Configurations : BaseState() {
     var autoSuggestExecutable by property(true)
     var workspaceFolders by enum(WorkspaceFolders.PROJECT_BASE)
     var targetedFileExtensions by string(listOf("py", "pyi").join())
-    var diagnosticMode by enum(DiagnosticMode.DEFAULT)
+    var diagnosticMode by enum(DiagnosticMode.OPEN_FILES_ONLY)
+    var autoSearchPaths by property(true)
 }