+++ src/main/kotlin/com/insyncwithfoo/pyrightls/server/Settings.kt
@@ -4,7 +4,8 @@ package com.insyncwithfoo.pyrightls.server
 internal data class Analysis(
     var logLevel: String? = null,
     var autoImportCompletions: Boolean? = null,
-    var diagnosticMode: String? = null
+    var diagnosticMode: String? = null,
+    var autoSearchPaths: Boolean = true
 )